Split shell tool executor tests
