+++ docs/DECISIONS.md
@@ -320,3 +320,4 @@
 - HTML review renderer coverage lives in `WorkspaceHTMLReviewRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Static review-pane markup, review file/action/hunk/line/comment rendering, and review action data attributes should stay beside review HTML while review-surface data projection remains in focused workspace surface tests until it has a separate builder suite.
 - HTML secondary-pane renderer coverage lives in `WorkspaceHTMLSecondaryPaneRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Extensions and Memories static HTML smoke tests should stay beside `WorkspaceHTMLSecondaryPaneRenderer`; browser HTML remains in browser-focused suites because the browser pane has a separate ownership boundary.
 - Workspace automation model orchestration lives in `WorkspaceModelAutomations.swift`, not in the main `WorkspaceModel.swift` body. The extension owns public automation scheduling/run/delete APIs and delegates pure mutations to `WorkspaceAutomationStateReducer` and draft creation to `WorkspaceAutomationRunner`; the main model owns only shared actor-isolated helpers for persistence, project context refresh, thread insertion, errors, and top-bar refresh.
+- Shell and SSH shell executor coverage lives in `ShellToolExecutorTests`, not in the mixed `ToolTests.swift` suite. Reusable temp directory, git repo, fake GitHub CLI, and fake SSH fixtures live in `ToolTestSupport.swift` so future focused tool suites can share setup without growing another broad catch-all test file.
